* FIX: Minor the details fixes

+++ code/engine.vc2008/xrGame/alife_monster_detail_path_manager.cpp
@@ -272,7 +272,7 @@ Fvector CALifeMonsterDetailPathManager::draw_level_position	() const
 	if (path().empty())
 		return						(object().get_object().Position());
 
-	u32								path_size = path().size();
+	size_t							path_size = path().size();
 	if (path_size == 1)
 		return						(object().get_object().Position());
